Fixed compiler warning.
git-svn-id: https://svn.opendtect.org/od/trunk@36774 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/gaussianprobdenfunc.cc
@@ -28,7 +28,7 @@ static inline float draw01Normal( const Stats::NormalRandGen& rgen )
 static inline float draw01Correlated( const Stats::NormalRandGen& rgen,
 					float othdraw, float cc )
 {
-    const float draw = rgen.get();
+    const float draw = (float) rgen.get();
     return cc * othdraw + Math::Sqrt(1 - cc*cc) * draw;
 }
